Initial implementation OSMNodes DikstraSP + Distance class
Still work in progress, can hopefully be done with in it sometime next week.
Distance class uses the Haversine formula to calculate distance between two points using lat and lon.

+++ Map of Denmark/app/src/main/java/bfst22/vector/Coordinates.java
@@ -25,19 +25,19 @@ public class Coordinates {
         }
     }
 
-    public float getMinX(){
+    public double getMinX(){
         return this.minX;
     }
 
-    public float getMinY(){
+    public double getMinY(){
         return this.minY;
     }
 
-    public float getMaxX(){
+    public double getMaxX(){
         return this.maxX;
     }
 
-    public float getMaxY(){
+    public double getMaxY(){
         return this.maxY;
     }
